Updated foreign keys so that it works

+++ database/migrations/2020_07_04_200000_create_users_table.php
@@ -23,7 +23,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->string('first_name')->nullable();
             $table->string('last_name')->nullable();
-            $table->integer('user_type_id')->unsigned();
+            $table->bigInteger('user_type_id')->unsigned();
             $table->rememberToken();
             $table->timestamps();
             $table->softDeletes();
